requirements.txt
opencv-python
numpy
PyYAML
pycocotools
tqdm
torch
torchvision
git+https://github.com/facebookresearch/segment-anything.git
